refactor(cli): give each floe command its own file
main.go 539 -> 187, which is now the root command, the shared flags and env
handling, peerOptions, connectedLine and the signal-handling entry point.
Everything else follows its command:

- send.go, the floe send command.
- receive.go, floe receive and the three flags only it registers.
- update.go, floe version and floe update, which are the same concern from a
  user's side: which build am I on, and how do I move.

The section banners the file used to carry are dropped, because the file name
now says what they said.

The risk in this split is not the code, it is the wiring. Cobra registers
flags from init functions, and there are three of them: main.go's root flags,
receive.go's, and update.go's. Go runs a package's init functions in file-name
order, and main.go still sorts before receive.go before update.go, so the
relative order is unchanged. That is an argument, not a proof, which is why
the proof is the help text.

go doc cannot see init ordering or flag registration at all, so this PR is
verified by diffing the full help output of every command against a binary
built from a worktree at the merge base. All five are byte-identical, and so
are the flag lists of receive and update specifically.

The plan for this PR also named the STUN filter dedupe. That is deliberately
not here: collapsing two copies into one REMOVES a symbol, so the diff would
no longer be a pure move, and this PR's whole value is that it is.

Test plan
- floe --help, send --help, receive --help, version --help, update --help:
  all five byte-identical to a binary built from the base worktree.
- The receive and update flag lists diffed on their own: identical.
- runSend, runReceive and runUpdate md5-identical to their base copies.
- go build, go vet and go test clean in cli/.
